Optimize project layout to avoid bad build performance
