+++ components/logins/Cargo.toml
@@ -34,10 +34,5 @@ features = ["sqlcipher", "limits", "unlock_notify"]
 
 [dev-dependencies]
 more-asserts = "0.2"
-env_logger = "0.7"
-prettytable-rs = "0.8"
-fxa-client = { path = "../fxa-client" }
-chrono = "0.4"
-clap = "2.33"
-cli-support = { path = "../support/cli" }
 tempdir = "0.3"
+env_logger = { version = "0.7", default-features = false }
