+++ components/support/ffi/Cargo.toml
@@ -25,8 +25,3 @@ lazy_static = "1.4"
 [dependencies.backtrace]
 optional = true
 version = "0.3.48"
-
-[dev-dependencies]
-rand = "0.7"
-rayon = "1.3"
-env_logger = "0.7"
