+++ components/sync15/Cargo.toml
@@ -29,4 +29,4 @@ thiserror = "1.0"
 anyhow = "1.0"
 
 [dev-dependencies]
-env_logger = "0.7"
+env_logger = { version = "0.7", default-features = false }
