+++ examples/fxa-client/src/devices-api.rs
