+++ examples/tabs-sync/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "example-tabs-sync"
+version = "0.1.0"
+authors = ["sync-team@mozilla.com"]
+license = "MPL-2.0"
+edition = "2018"
+publish = false
+
+[[example]]
+name = "tabs-sync"
+path = "src/tabs-sync.rs"
+
+[dev-dependencies]
+tabs = { path = "../../components/tabs" }
+serde_json = "1"
+log = "0.4"
+anyhow = "1.0"
+clipboard = "0.5"
+clap = "2.33"
+cli-support = { path = "../../components/support/cli" }
+viaduct-reqwest = { path = "../../components/support/viaduct-reqwest" }
