+++ testing/separated/README.md
@@ -1,0 +1,4 @@
+# Separated tests
+
+These are pulled out for compile-time reasons. See `docs/test-faster.md` for an
+in-depth explanation and recommended guide for how to add things.
