+++ testing/separated/ffi-support-extra-test/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "ffi-support-extra-test"
+version = "0.1.0"
+authors = ["sync-team@mozilla.com"]
+license = "MPL-2.0"
+edition = "2018"
+publish = false
+
+[[test]]
+path = "src/test.rs"
+name = "ffi-support-extra"
+
+[dev-dependencies]
+ffi-support = { version = "0.4", path = "../../../components/support/ffi" }
+rayon = "1.3"
+env_logger = { version = "0.7", default-features = false }
+log = "0.4"
+rand = "0.7"
+
