+++ testing/separated/places-bench/fixtures/README.md
@@ -1,0 +1,2 @@
+See components/places/docs/dummy_urls.md for background on how this file was
+collected.
