+++ testing/separated/places-tests/src/fennec_bookmarks_schema.sql
