+++ testing/separated/places-tests/src/fennec_history_schema.sql
